+++ packages/react/index.js
@@ -8,7 +8,7 @@ export { Badge } from "./Badge.jsx";
 export { Avatar, Link, Tag, Banner, CommandPalette } from "./Misc.jsx";
 
 // Formulário
-export { Field, Input, PhoneInput } from "./Field.jsx";
+export { Field, Input, PhoneInput, SU_COUNTRIES, suDetectCountry } from "./Field.jsx";
 export { FormGrid } from "./FormGrid.jsx";
 export { Select, Checkbox, CheckGroup, Radio, Switch, SegmentedControl } from "./Controls.jsx";
 export { NumericInput, TextArea, Combobox, FileUpload, DatePicker } from "./Inputs.jsx";